requirements.txt
# Discord Bot with LangChain, LangMem, and MCP Server Tools
discord.py>=2.0.0
python-dotenv>=0.21.0
langchain>=0.1.0
langchain-community>=0.0.13
langchain-mistralai>=0.0.1
langchain-google-genai>=0.0.3
langchain-mcp-adapters>=0.0.1
langmem>=0.0.1
langgraph>=0.0.20
chromadb>=0.4.22  # Added for persistent memory
langgraph-checkpoint-sqlite
aiosqlite>=0.20.0